added modes changed a few minor things, some experimentation

+++ Util.pde
@@ -1,8 +1,8 @@
 
 static class Util{
 
-    static double sigmoid(float x){
-        return 1.0 / (1.0 + Math.exp(-x)); 
+    static float sigmoid(float x){
+        return (float)(1.0 / (1.0 + Math.exp(-x))); 
     }
 
     static float primeSigmoid(float x){
